Revert "Load configured Active Storage plugins during boot"
