+++ activestorage/lib/active_storage.rb
@@ -49,8 +49,6 @@ module ActiveStorage
   mattr_accessor :verifier
   mattr_accessor :variant_processor, default: :mini_magick
 
-  mattr_accessor :variant_transformer
-
   mattr_accessor :queues, default: {}
 
   mattr_accessor :previewers, default: []
@@ -378,7 +376,5 @@ module ActiveStorage
 
     autoload :Transformer
     autoload :ImageProcessingTransformer
-    autoload :Vips
-    autoload :ImageMagick
   end
 end
