+++ activestorage/lib/active_storage/analyzer/image_analyzer/vips.rb
@@ -1,18 +1,22 @@
 # frozen_string_literal: true
 
-begin
-  gem "ruby-vips"
-  require "ruby-vips"
-rescue LoadError => error
-  raise error unless error.message.include?("ruby-vips")
-end
-
 module ActiveStorage
   # This analyzer relies on the third-party {ruby-vips}[https://github.com/libvips/ruby-vips] gem. Ruby-vips requires
   # the {libvips}[https://libvips.github.io/libvips/] system library.
   class Analyzer::ImageAnalyzer::Vips < Analyzer::ImageAnalyzer
+    def self.accept?(blob)
+      super && ActiveStorage.variant_processor == :vips
+    end
+
     private
       def read_image
+        begin
+          require "ruby-vips"
+        rescue LoadError
+          logger.info "Skipping image analysis because the ruby-vips gem isn't installed"
+          return {}
+        end
+
         download_blob_to_tempfile do |file|
           image = instrument("vips") do
             # ruby-vips will raise Vips::Error if it can't find an appropriate loader for the file
@@ -31,9 +35,6 @@ module ActiveStorage
           logger.error "Skipping image analysis due to a Vips error: #{error.message}"
           {}
         end
-      rescue ::Vips::Error => error
-        logger.error "Skipping image analysis due to an Vips error: #{error.message}"
-        {}
       end
 
       ROTATIONS = /Right-top|Left-bottom|Top-right|Bottom-left/
