--- activestorage/lib/active_storage/transformers/vips.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module ActiveStorage
-  module Transformers
-    class Vips < ImageProcessingTransformer
-      def processor
-        ImageProcessing::Vips
-      end
-    end
-  end
-end
