+++ activestorage/test/models/variant_test.rb
@@ -295,21 +295,11 @@ class ActiveStorage::VariantTest < ActiveSupport::TestCase
 
   private
     def process_variants_with(processor)
-      previous_transformer = ActiveStorage.variant_transformer
-      ActiveStorage.variant_transformer =
-        case processor
-        when :vips
-          ActiveStorage::Transformers::Vips
-        when :mini_magick
-          ActiveStorage::Transformers::ImageMagick
-        else
-          raise "#{processor.inspect} is not a valid image transformer"
-        end
-
+      previous_processor, ActiveStorage.variant_processor = ActiveStorage.variant_processor, processor
       yield
     rescue LoadError
       ENV["BUILDKITE"] ? raise : skip("Variant processor #{processor.inspect} is not installed")
     ensure
-      ActiveStorage.variant_transformer = previous_transformer
+      ActiveStorage.variant_processor = previous_processor
     end
 end
